- use first footfall journal entry, even when exiting an SRV - make sure a folder exists before writing a file in it - Force handling VR adjustments onto the UI thread - Fix "The given key '$Codex_RegionName_18;' was not present in the dictionary." - Avoid crash in CreateParams by overriding: GetPreferredSize - Make adjusting locations in VR a little smoother

+++ SrvSurvey/forms/FormSetKeyChord.cs
@@ -20,7 +20,7 @@ namespace SrvSurvey
             // force form width as it doesn't happen correctly by itself
             this.Width = flowButtons.Right + (flowButtons.Left * 2) + 4;
 
-            KeyboardHook.buttonsPressed += (bool hook, string chord, short analog) =>
+            KeyboardHook.buttonsPressed += (bool hook, string chord, int analog) =>
             {
                 btnAccept.Enabled = hook;
                 this.textChord.Text = chord;
